# mail: use stdlib instead of external dependency