--- a/go.mod
+++ b/go.mod
@@ -3,9 +3,6 @@ module github.com/emersion/go-message
 go 1.13
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/emersion/go-textwrapper v0.0.0-20200911093747-65d896831594
-	github.com/martinlindhe/base36 v1.1.0
-	github.com/stretchr/testify v1.3.0 // indirect
 	golang.org/x/text v0.3.5-0.20201125200606-c27b9fd57aec
 )