--- a/mail/header.go
+++ b/mail/header.go
@@ -1,7 +1,6 @@
 package mail
 
 import (
-	"bytes"
 	"crypto/rand"
 	"encoding/binary"
 	"errors"
@@ -9,12 +8,12 @@ import (
 	"net/mail"
 	"os"
 	"regexp"
+	"strconv"
 	"strings"
 	"time"
 	"unicode/utf8"
 
 	"github.com/emersion/go-message"
-	"github.com/martinlindhe/base36"
 )
 
 const dateLayout = "Mon, 02 Jan 2006 15:04:05 -0700"
@@ -308,22 +307,26 @@ func (h *Header) MsgIDList(key string) ([]string, error) {
 // informational draft "Recommendations for generating Message IDs", for lack
 // of a better authoritative source.
 func (h *Header) GenerateMessageID() error {
-	now := bytes.NewBuffer(make([]byte, 0, 8))
-	binary.Write(now, binary.BigEndian, time.Now().UnixNano())
+	now := uint64(time.Now().UnixNano())
 
-	nonce := make([]byte, 8)
-	if _, err := rand.Read(nonce); err != nil {
+	nonceByte := make([]byte, 8)
+	if _, err := rand.Read(nonceByte); err != nil {
 		return err
 	}
+	nonce := binary.BigEndian.Uint64(nonceByte)
 
 	hostname, err := os.Hostname()
 	if err != nil {
 		return err
 	}
 
-	msgID := fmt.Sprintf("%s.%s@%s", base36.EncodeBytes(now.Bytes()), base36.EncodeBytes(nonce), hostname)
+	msgID := fmt.Sprintf("%s.%s@%s", base36(now), base36(nonce), hostname)
 	h.SetMessageID(msgID)
 	return nil
+}
+
+func base36(input uint64) string {
+	return strings.ToUpper(strconv.FormatUint(input, 36))
 }
 
 // SetMessageID sets the Message-ID field. id is the message identifier,