I'm not sure if there is a specific reason for doing it as it currently is? Anyhow, this pull request removes this otherwise unused external dependency and use strconv to base36 encode. This has the side effect of reducing allocations as well as improving execution times.  Output should be comparable;

```
Old: C9EDU8EPQBGY.1OPA7WEUKTFIY@hostname
New: C9EDU7U5NQNZ.1T2B0PIFP4V88@hostname
Old: C9EE4CRAIMU9.35K02SYKO9VWW@hostname
New: C9EE4C6I3XNX.2V6K31R712RT@hostname
```

And benchmark diff using `benchstat` when calling the `GenerateMessageID` function;
```
name                      old time/op    new time/op    delta
Header_GenerateMessageID    8.71µs ± 5%    6.59µs ±10%  -24.29%  (p=0.000 n=10+10)

name                      old alloc/op   new alloc/op   delta
Header_GenerateMessageID      606B ± 0%      275B ± 0%  -54.58%  (p=0.000 n=10+10)

name                      old allocs/op  new allocs/op  delta
Header_GenerateMessageID      45.0 ± 0%      13.0 ± 0%  -71.11%  (p=0.000 n=10+10)
```